fez_dev/backend/requirements.txt
Django == 3.2.4
uwsgi == 2.0.19.1
PyMySQL == 1.0.2
mysqlclient == 2.0.3
cryptography == 3.4.7
djangorestframework == 3.12.4
djangorestframework-simplejwt == 4.7.1
django-cors-headers == 3.7.0
django-seed== 0.2.2
unicodecsv == 0.14.1